% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_parents.R
\name{drop_parents}
\alias{drop_parents}
\title{Drop Parents and F1 Individuals from a Data Frame}
\usage{
drop_parents(y, parent1 = "P1", parent2 = "P2", F1 = NULL)
}
\arguments{
\item{y}{A data frame where individuals (including parents and F1s) are columns.}

\item{parent1}{Character. The name of the first parent column to be removed. Default is \code{"P1"}.}

\item{parent2}{Character. The name of the second parent column to be removed. Default is \code{"P2"}.}

\item{F1}{(Optional) Character. The name of the F1 hybrid column to be removed. Default is \code{NULL}.}
}
\value{
A data frame with the specified parent and F1 columns removed.
}
\description{
This helper function removes specified parental and F1 columns from a genotype
dataset.
}
\details{
\itemize{
\item \code{y} must be coercible to a data frame.
\item Checks if the specified columns exist before attempting to drop them.
\item Only removes columns that are present in the data frame.
\item Preserves all other columns and their structure.
\item If multiple F1s are present they can be declared using the \code{c()} function
}
}
\examples{
# Example dataset
geno_data <- data.frame(
  Marker1 = c(0, 1, 2),
  P1 = c(0, 0, 0),
  P2 = c(2, 2, 2),
  F1 = c(1, 1, 1),
  Ind1 = c(0, 1, 2),
  Ind2 = c(2, 0, 1)
)

# Drop parents P1 and P2
filtered_data <- drop_parents(geno_data)
print(filtered_data)

# Drop parents and F1
filtered_data_f1 <- drop_parents(geno_data, F1 = "F1")
print(filtered_data_f1)

# Example dataset2
geno_data2 <- data.frame(
  Marker1 = c(0, 1, 2),
  P1 = c(0, 0, 0),
  P2 = c(2, 2, 2),
  F1a = c(1, 1, 1),
  F1b = c(1, 1, 1),
  Ind1 = c(0, 1, 2),
  Ind2 = c(2, 0, 1)
)


# Drop parents and multiple F1s
filtered_data_f1 <- drop_parents(geno_data, F1 = c("F1a", "F1b"))
print(filtered_data_f1)

}
