% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_plot.R
\name{effect_plot}
\alias{effect_plot}
\title{Powerful Effect Plots}
\usage{
effect_plot(
  effects_df,
  marker_name,
  trait_name,
  genotype_levels = c("A", "H", "B"),
  flip = TRUE,
  trait_label = NULL
)
}
\arguments{
\item{effects_df}{A data frame containing at least the columns for marker and trait of interest.
Could work with the output of \code{format_qtl_input}with a few modifications (first two rows removed).}

\item{marker_name}{A character string specifying the name of the marker column in \code{effects_df}.}

\item{trait_name}{A character string specifying the name of the trait column in \code{effects_df}.}

\item{genotype_levels}{A character vector indicating the levels (genotype categories) for the marker. Default is \code{c("A", "H", "B")}.}

\item{flip}{Logical; if \code{TRUE}, the plot is displayed with flipped coordinates (horizontal layout). Default is \code{TRUE}.}

\item{trait_label}{Optional character string to use as a y-axis label. If \code{NULL}, the trait name will be used.}
}
\value{
A list with three elements:
\describe{
\item{\code{[[1]]}}{A \code{ggplot2} object showing the distribution of trait values for each genotype at the given marker.}
\item{\code{[[2]]}}{A data frame with summary statistics (median, max, min, count, standard deviation) by genotype.}
\item{\code{[[3]]}}{A data frame with population-level summary statistics (mean, median, sd, max) for the trait.}
}
}
\description{
This function generates powerful effect plots to show the relationship between genotype classes at a a marker (QTL) on a given trait.
\strong{Works on polyploids}
The function computes summary statistics for each genotype class and overlays them on the plot.
Optionally, it can flip the coordinate axes.
Shows the distributions and measures of spread of individuals and not simply a boxplot.
}
\details{
The function performs the following steps:
\itemize{
\item Checks for presence of specified columns.
\item Calculates summary statistics by genotype.
\item Constructs a half-eye plot with boxplot overlays and annotations.
}
It uses \code{ggplot2} for plotting and \code{ggdist} for the half-eye visualization.
}
\note{
The half-eye visualization is inspired in work by \href{https://www.cedricscherer.com/2021/06/06/visualizing-distributions-with-raincloud-plots-and-how-to-create-them-with-ggplot2/}{Cedric Scherer}
}
\examples{
\dontrun{
effect_plot(effects_df = my_data,
                marker_name = "SNP_1",
                trait_name = "PlantHeight",
                genotype_levels = c("A", "H", "B"),
                flip = TRUE,
                trait_label = "Height (cm)")

 #Works for polyploids if genotype levels are specified and match the `effects_df`.

effect_plot(effects_df = my_data,
                marker_name = "SNP_1",
                trait_name = "PlantHeight",
                genotype_levels = c("0", "1", "2", "3", "4"), # or
                  c("AAAA", "AAAB", "AABB", "ABBB", "BBBB")
                flip = TRUE,
                trait_label = "Height (cm)")
}
}
