% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geno_by_parents.R
\name{filter_geno_by_parents}
\alias{filter_geno_by_parents}
\title{Filter Genotype Matrix Based on Parental Genotype}
\usage{
filter_geno_by_parents(geno, parent1, parent2)
}
\arguments{
\item{geno}{A genotype matrix or data frame where markers are rows and individuals are columns.
Must be coercible to a data frame.}

\item{parent1}{Character. The name of the column representing the first parent.}

\item{parent2}{Character. The name of the column representing the second parent.}
}
\value{
A filtered data frame containing only markers that are homozygous polymorphic
between the two parents.
}
\description{
\code{filter_geno_by_parents} filters a genotype matrix to retain only homozygous polymorphic markers
between two specified parents. It removes markers that do not meet the homozygous
polymorphism criteria (\code{P1 = 0} & \code{P2 = 2} or \code{P1 = 2} & \code{P2 = 0}).
}
\details{
\itemize{
\item Retains markers where \code{parent1} is \code{0} and \code{parent2} is \code{2} (\verb{A × B})
or \code{parent1} is \code{2} and \code{parent2} is \code{0} (\verb{B × A}). See the methods section
of Braun et al. 2017 for  more info on marker types. \href{https://acsess.onlinelibrary.wiley.com/doi/pdf/10.3835/plantgenome2016.10.0110}{The Plant Genome Vol. 10 No. 3.}
}
}
\examples{
# Example genotype matrix
geno_data <- data.frame(
  Marker1 = c(0, 1, 2, 0, 2),
  Marker2 = c(2, 0, 2, 1, 0),
  Parent1 = c(0, 2, 2, 0, 2),
  Parent2 = c(2, 0, 0, 2, 0)
)

# Filter markers based on parents
filtered_geno <- filter_geno_by_parents(geno_data, "Parent1", "Parent2")
print(filtered_geno)

}
