% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_qtl_input.R
\name{format_qtl_input}
\alias{format_qtl_input}
\title{Format Input Data for R/qtl package}
\usage{
format_qtl_input(geno, map, pheno, numeric = TRUE)
}
\arguments{
\item{geno}{A dataframe containing genotype data with markers as rows and individuals as columns.}

\item{map}{A dataframe containing the genetic map with columns: marker, chrom, and position.}

\item{pheno}{A dataframe containing phenotype data with an ID column corresponding to individual/sample names.}

\item{numeric}{A logical value (TRUE or FALSE) indicating whether to convert dosages to R/qtl ABH format (default = TRUE).}
}
\value{
A dataframe combining genotype, map, and phenotype data in \href{https://cran.r-project.org/web/packages/qtl/index.html}{R/qtl} format  and ready for \link[qtl]{read.cross} after saving it as csv.
}
\description{
This function is part of the heart of the package. \code{format_qtl_input} formats genotype, map, and phenotype data for QTL analysis in \href{https://cran.r-project.org/web/packages/qtl/index.html}{R/qtl}.
It allows the user to specify whether the genotype data should be converted from dosage to ABH format or used as is.
}
\note{
Formats input data for the read.cross function of \href{https://rqtl.org/}{R/qtl}.
}
\examples{
\dontrun{
# Example usage:
# result <- format_qtl_input(geno2, map2, pheno2, numeric = TRUE)
}
}
\seealso{
\code{\link[=write.csv]{write.csv()}}, \code{\link[=read.cross]{read.cross()}}
}
