% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOD_int.R
\name{get_LOD_int}
\alias{get_LOD_int}
\title{Get LOD Support Intervals for QTL Peaks (Single or Multiple Traits)}
\usage{
get_LOD_int(cross_obj, model_obj, results_obj, trait = NULL, drop = 1.5)
}
\arguments{
\item{cross_obj}{A \code{cross} object from the \pkg{qtl} package. Usually the output of \link[qtl]{read.cross} or \link[qtl]{calc.genoprob}.}

\item{model_obj}{Either a scanone/CIM object (class "scanone") or a named list of such objects. Usuall the output of \code{cim} or \code{scanone} from \pkg{qtl}.}

\item{results_obj}{Either a data frame of QTL results (with "chr" and "pos") or a named list of such data frames.}

\item{trait}{Optional. Name of the trait to analyze. Required if model_obj and results_obj are lists (i.e., when working with multiple traits).}

\item{drop}{Numeric. LOD drop to define the support interval (default = 1.5).}
}
\value{
A named list of QTL interval summaries, flanking markers, peak marker (pseudo marker or physical marker), and usable marker (closest physical marker if peak is on a pseudo marker).
}
\description{
Computes LOD support intervals for detected QTLs from \link[qtl]{scanone} or \link[qtl]{cim}.
}
\details{
Works with either a single model and qtl results or a list of models and results for multiple traits.
}
\examples{
\dontrun{
# Multiple traits:
# For the length-width trait
lw_result_pop1 <- get_LOD_int(
  cross_obj = M1,
  model_obj = cim_qtl_results1,
  results_obj = results_pop1,
  trait = "length_width_ratio",
  drop = 1.5
)

# Single traits:
mod <- cim_qtl_results1[["length_width_ratio"]]
res <- results_pop1[["length_width_ratio"]]

single <- get_LOD_int(
  cross_obj = M1,
  model_obj = mod,
  results_obj = res
)
}

}
