% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD_decay_plot.R
\name{LD_decay_plot}
\alias{LD_decay_plot}
\title{Plot LD vs distance}
\usage{
LD_decay_plot(
  data,
  map,
  max.pair = 10000,
  max.loci = NULL,
  position = "bp",
  r2_threshold = 0.2,
  xlim_range = c(0, 20),
  show_vline = TRUE,
  show_hline = TRUE,
  seed = 123
)
}
\arguments{
\item{data}{Genotype matrix (markers as columns, individuals as rows). Rownames have the marker names that matcho those of the marker column in the \code{map}.}

\item{map}{Data frame with columns 'chrom', 'marker', and 'position'}

\item{max.pair}{Maximum number of r² pairs for the model (default: 1e4)}

\item{max.loci}{Maximum number of markers to use per chromosome (default: NULL)}

\item{position}{"bp" or "Mb" (default: "bp")}

\item{r2_threshold}{r² value for decay distance calculation (default: 0.2),}

\item{xlim_range}{Zoom in range on the x axis, default is c(0, 20). Set to NULL to see entire plot.}

\item{show_vline}{show a vertical line intersecting the x-axis at the the half-decay distance (default, TRUE)}

\item{show_hline}{show a horizontal line intersecting the y-axis at the selected r² threshold (default, TRUE)}

\item{seed}{Optional integer seed to make the random sampling reproducible. Default 123.}
}
\value{
A list containing:
\item{plot}{ggplot2 object with points and fitted curve (or NULL if fit fails)}
\item{half_decay_dist}{Distance where r² reaches the specified threshold (or NA if not estimable)}
\item{model}{Fitted nls object (or NULL if fitting failed)}
}
\description{
Plot LD vs distance using an asymptotic regression model.
}
\details{
An asymptotic regression model is fit using \link[stats]{SSasymp}.
The distance where r² reaches a specified threshold (e.g., 0.1 or 0.2) is calculated and returned.
This function calculates r² by chromosome and  follows the general logic of \href{https://github.com/jendelman/MapRtools/blob/master/R/plot_LD.R}{MapRtools::LD.plot} but uses the function
\link[stats]{SSasymp} to model the non-linear LD decay. This function will give you a
quick and simple estimate of LD. For more sophisticated LD functions see \href{https://github.com/dcgerard}{David Gerard} \href{https://cran.r-project.org/web/packages/ldsep/vignettes/fast.html}{ldfast()}

It processes each chromosome separately to compute all pairwise r²
values within chromosomes, then it combines the results across all
chromosomes into a single dataset, and fits one global LD decay model
to the pooled data. No cross-chromosome marker pairs are considered.

If the user is interested in LD per chromosome, provide a map and data file with only
data pertaining to the chromosome of interest.
}
\note{
If the number of rows (number of marker pairs) is larger than max.pair (default is 1e4), we proceed to random sampling.
}
