% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install_cran.R
\name{load_or_install_cran}
\alias{load_or_install_cran}
\title{Load or Install a CRAN Package}
\usage{
load_or_install_cran(pkg)
}
\arguments{
\item{pkg}{A character string specifying the name of the package to be loaded or installed.}
}
\value{
This function does not return a value. It is called for its side effect of loading or installing the package.
}
\description{
This function checks if a specified package is installed. If the package is
not installed, it installs it from CRAN, then loads it into the R session.
}
\details{
The function first checks if the package is available in the current session using \code{requireNamespace}.
If the package is not installed, it installs it using \code{install.packages} and then loads it using \code{library}.
}
\examples{
\dontrun{
load_or_install_cran("ggplot2")
load_or_install_cran("dplyr")
}

}
