% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_install.R
\name{load_or_install_github}
\alias{load_or_install_github}
\title{Load or install a GitHub package}
\usage{
load_or_install_github(pkg, repo)
}
\arguments{
\item{pkg}{Character. Package name.}

\item{repo}{Character. GitHub repo string in the form "user/repo".}
}
\value{
Loads the package. Installs it from GitHub if needed.
}
\description{
Load or install a GitHub package
}
