% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_and_plot.R
\name{order_and_plot}
\alias{order_and_plot}
\title{Optimize Marker Order}
\usage{
order_and_plot(
  trimmed_geno,
  pop.type = "F2",
  CHR = NULL,
  n.iter = 6,
  prop = NULL
)
}
\arguments{
\item{trimmed_geno}{A genotype matrix where:
\itemize{
\item Rows represent markers.
\item Columns represent individuals.
It could be the output of \code{trim_LG} or any genotype matrix.
}}

\item{pop.type}{Character. The population type used for ordering markers.
Default is \code{"F2"}. \code{order_and_plot} would work with any of the following "DH","BC","F2","S1","RIL.self","RIL.sib". Based on \code{MapRtools::MLEL()}}

\item{CHR}{Character (optional). The chromosome identifier for labeling plots.}

\item{n.iter}{Integer. The number of iterations for marker ordering. Default is \code{6}.}

\item{prop}{Numeric. The proportion of individuals to include in the genotype plot.
Must be between \code{0} and \code{1}. Default is \code{0.20} (20\% of individuals). If all individuals are included the visualization quality drops.}
}
\value{
A list containing:
\itemize{
\item \code{"original_geno"}: The original genotype matrix.
\item \code{"ordered_geno"}: The optimized genotype matrix with markers reordered.
\item \code{"order_plot"}: A \code{ggplot2} object displaying the original vs. optimized marker order.
\item \code{"haplotype_plot"}: A \code{ggplot2} object showing the genotype haplotype plot.
}
}
\description{
\code{order_and_plot} optimizes the marker order in a (trimmed) genotype matrix using
recombination frequency (RF) matrices. It performs multiple iterations of
marker ordering, selects the best order based on the smallest Sum of Adjusted
Recombination Fractions (SARF), and generates visualizations of marker order. The function uses
\code{MLEL()} and \code{order_markers}() from Professor Jeffrey Endelman's \href{https://github.com/jendelman/MapRtools}{MapRtools} and streamlines the process.
This workflow should be easier on the user when dealing with multiple chromosomes.
\href{https://github.com/jendelman/MapRtools}{MapRtools}.
}
\details{
\itemize{
\item Computes an RF matrix using \code{MLEL()}, then orders markers iteratively.
\item Runs \code{order_markers()} \code{n.iter} times and selects the order with the smallest SARF.
\item \code{order_markers()} Order markers by solving the traveling salesperson problem.
\item Plots the original vs. optimized marker order.
\item Generates a genotype haplotype plot for a subset of individuals (\code{prop}).
\item If \code{CHR} is provided, adds a chromosome label to the plots.
}
}
\note{
This function was refined with assistance from ChatGPT to improve clarity,
efficiency, and visualization formatting.
}
