% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot Genome-wide Coverage}
\usage{
plot_coverage(map, limits = NULL, customize = FALSE)
}
\arguments{
\item{map}{A data frame with two columns:
\itemize{
\item \code{"chrom"}: Chromosome identifier (e.g., \verb{"CHR1", "CHR2", ..."}).
\item \code{"position"}: The genomic position of markers (in base pairs).
}}

\item{limits}{(Optional) A data frame specifying the maximum position for each chromosome.
If \code{NULL} (default), the function computes chromosome limits from \code{map}.}

\item{customize}{Logical. If \code{TRUE}, applies additional visual customizations.
Default is \code{FALSE}.}
}
\value{
A \code{ggplot2} object representing genome-wide coverage.
}
\description{
Generates a genome-wide coverage plot, displaying the positions
of markers across chromosomes. It is a customized version of \code{plot_coverage}
from \strong{MapRtools} and allows additional aesthetic modifications.
}
\details{
\itemize{
\item Converts genomic positions from base pairs to megabases (Mb).
\item If \code{limits} are not provided, the function calculates the maximum position
for each chromosome from \code{map}.
\item Orders chromosomes and positions correctly for visualization.
\item Uses \code{geom_segment()} to generate a chromosome-wide coverage plot.
\item When \code{customize = TRUE}, applies a minimalistic theme for enhanced visualization.
}
}
\note{
Inspired by and \strong{reliant} on functions by Professor Jeffrey B. Endelman's MapRtools
}
\examples{
# Example dataset
map_data <- data.frame(
  chrom = c("CHR1", "CHR1", "CHR2", "CHR2", "CHR3"),
  position = c(500000, 1200000, 800000, 1600000, 2000000)
)

# Basic coverage plot
plot_coverage(map_data)

# Coverage plot with custom aesthetics
plot_coverage(map_data, customize = TRUE)


# With user-defined limits
limits_data <- data.frame(
 chrom = c("CHR1", "CHR2", "CHR3"),
 max_position = c(1500000, 2000000, 2500000)
)

plot_coverage(map_data, limits = limits_data)


}
