% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_effect_plot.R
\name{simple_effect_plot}
\alias{simple_effect_plot}
\title{Simple Effect Plot}
\usage{
simple_effect_plot(
  effects_df,
  marker_name,
  trait_name,
  genotype_levels = c("A", "H", "B"),
  annotate = TRUE,
  flip = FALSE,
  trait_label = NULL
)
}
\arguments{
\item{effects_df}{A data frame containing at least the columns for marker genotypes and trait values. Usually the output of \code{format_qtl_input}with a few modifications (first two rows removed).}

\item{marker_name}{A character string specifying the name of the marker column in \code{effects_df}.}

\item{trait_name}{A character string specifying the name of the trait column in \code{effects_df}.}

\item{genotype_levels}{A character vector indicating the expected genotype categories (factor levels). Default is \code{c("A", "H", "B")}.}

\item{annotate}{Logical; if \code{TRUE}, the plot includes annotations of median values and sample sizes per genotype. Default is \code{TRUE}.}

\item{flip}{Logical; if \code{TRUE}, the coordinates of the plot are flipped (horizontal layout). Default is \code{FALSE}.}

\item{trait_label}{Optional character string to use as the y-axis label. If \code{NULL}, the trait name is used.}
}
\value{
A \code{ggplot2} object displaying the distribution of trait values across genotype classes (alleles) at the specified marker.
}
\description{
Generates a clean and simple effect plot. Shows the relationship between genotype classes at a marker on a trait.
Optionally includes annotation of medians and sample sizes, and allows flipping of axes.
\strong{Works on polyploids.}
}
\details{
The function performs the following:
\itemize{
\item Verifies the presence of marker and trait columns in the input data frame.
\item Removes rows with missing genotype data.
\item Converts the marker column to a factor with specified genotype levels.
\item Creates a boxplot with optional annotations of median values and counts.
\item Optionally flips the plot horizontally.
}
}
\note{
Inspired by  Fig. 2  \href{https://doi.org/10.1007/s00122-022-04159-z}{Caraza-Harter & Endelman}
}
\examples{
\dontrun{
simple_effect_plot(effects_df = my_data,
                        marker_name = "SNP_42",
                        trait_name = "RootWeight",
                        genotype_levels = c("A", "H", "B"),
                        annotate = TRUE,
                        flip = FALSE,
                        trait_label = "Weight (g)")

# Works for polyploids if genotype levels are specified and match the `effects_df`.

simple_effect_plot(effects_df = my_data,
                marker_name = "SNP_1",
                trait_name = "PlantHeight",
                genotype_levels = c("0", "1", "2", "3", "4"),
                # or  c("AAAA", "AAAB", "AABB", "ABBB", "BBBB")
                flip = TRUE,

}

}
