% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_geno.R
\docType{data}
\name{simulated_geno}
\alias{simulated_geno}
\title{Simulated Simple Genotype Matrix Dataset}
\format{
A data frame with 6 rows and 5 variables:
\describe{
\item{Parent1}{Parent1 genotype in dosage of ALT allele (numeric)}
\item{Parent2}{Parent2 genotype in dosage of ALT allele (numeric)}
\item{F2_1}{Progeny genotype in dosage of ALT allele (numeric)}
\item{F2_2}{Genotype of F2 individual 2}
\item{F2_3}{Genotype of F2 individual 3}
}
}
\source{
Simulated data for internal use.
}
\usage{
simulated_geno
}
\description{
This dataset contains a sample genotype matrix for demonstration of \code{geneticMapper} functions.
}
\keyword{datasets}
