% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_LG.R
\name{trim_LG}
\alias{trim_LG}
\title{Trim Linkage Groups Based on LOD Thresholds}
\usage{
trim_LG(
  chromosome,
  map,
  geno,
  pop_type = "F2",
  drop_outliers = TRUE,
  n_cores = NULL,
  interactive = TRUE,
  initial_LOD = NULL,
  end_LOD = NULL,
  step_LOD = NULL,
  selected_thresh = NULL
)
}
\arguments{
\item{chromosome}{Character. The chromosome ID to be processed.}

\item{map}{A data frame with at least columns \code{"marker"}, \code{"chrom"}, and \code{"position"}.}

\item{geno}{A genotype matrix (rows = markers, columns = individuals). Preferably binned.}

\item{pop_type}{Character. Population type for LOD estimation. Default is \code{"F2"}.}

\item{drop_outliers}{Logical. Whether to remove outliers based on haplotype frequency. Default is \code{TRUE}.}

\item{n_cores}{Integer. Number of cores to use. If \code{NULL}, uses maximum available minus one.}

\item{interactive}{Logical. If \code{TRUE}, asks user for LOD thresholds. If \code{FALSE}, requires threshold arguments.}

\item{initial_LOD}{Numeric. Initial LOD value (required if \code{interactive = FALSE}).}

\item{end_LOD}{Numeric. Final LOD value (required if \code{interactive = FALSE}).}

\item{step_LOD}{Numeric. Step size for LOD sequence (required if \code{interactive = FALSE}).}

\item{selected_thresh}{Numeric. Final LOD threshold to use (required if \code{interactive = FALSE}).}
}
\value{
A list with filtered genotypes, final map, plots, and LOD settings.
}
\description{
\code{trim_LG} is an interactive or scripted function that helps the user filter markers within a specified chromosome
based on linkage group (LG) assignment using LOD score thresholds. The user can interactively choose LOD thresholds
or provide them via arguments, visualize haplotype frequency before and after filtering, and remove outliers.
}
\details{
Inspired by functions by Professor Jeffrey B. Endelman’s \href{https://github.com/jendelman/MapRtools}{MapRtools}.
}
