% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_reconstruction.R
\name{haplotype_reconstruction}
\alias{haplotype_reconstruction}
\title{F2 diploid population Haplotype Reconstruction}
\usage{
haplotype_reconstruction(geno_matrix, error_rate = 0.05, r = 0.01)
}
\arguments{
\item{geno_matrix}{A numeric genotype matrix where:
\itemize{
\item Rows represent genetic markers.
\item Columns represent progeny (individuals).
\item Values are expected to be \code{0}, \code{1}, \code{2}, or \code{NA} for missing data.
}}

\item{error_rate}{Numeric. The assumed genotyping error rate. Default is \code{0.05} (5\% error rate).}

\item{r}{Numeric. The recombination rate used in the transition probability matrix (\code{T.mat}).
Default is \code{0.01}. The user can use the \code{MLEL} function from \href{https://github.com/jendelman/MapRtools}{MapRtools}
to get an estimate of recombination frequency with the adjacent argument set to \code{TRUE}
\code{result <- MLEL(geno = geno_matrix, pop.type = "f2", LOD = FALSE, adjacent = TRUE)}
In this case the geno_matrix should be of only one chromosome. \code{r} could be estimated as \code{mean(result$value, na.rm = TRUE)}}
}
\value{
A genotype matrix with inferred haplotypes based on HMM correction.
}
\description{
This function performs haplotype reconstruction using a Hidden Markov Model (HMM).
It applies the Viterbi algorithm to infer the most likely sequence of true genotypic
states, accounting for genotyping errors and missing data.
}
\details{
\itemize{
\item Estimates the missing data rate in the input genotype matrix.
\item Initializes an HMM using:
\itemize{
\item Three hidden states (\code{"0"}, \code{"1"}, \code{"2"}).
\item Four observable symbols (\code{"0"}, \code{"1"}, \code{"2"}, \code{"NA"} for missing data).
\item Transition probabilities generated via \code{T.mat(r = r)}, using the user-defined recombination rate.
\item Emission probabilities computed using \code{E.mat(error = error_rate, missing = missing)}.
}
}
}
\note{
Works only for F2 populations on experimental crosses. Based on the 615 Genetic Mapping Class notes by
Prof. Jeffrey Endelman. Spring 2021.
}
\examples{
\dontrun{
# Example genotype matrix with missing values
geno_data <- matrix(c(0, 1, NA, 2, 0, 1, NA, 2, NA, NA, 0, 1),
                    nrow = 4, ncol = 3,
                    dimnames = list(c("Marker1", "Marker2", "Marker3", "Marker4"),
                                    c("Ind1", "Ind2", "Ind3")))

# Perform haplotype reconstruction with default parameters
reconstructed_geno <- haplotype_reconstruction(geno_data)
print(reconstructed_geno)

# Perform haplotype reconstruction with a modified recombination rate
reconstructed_geno_custom_r <- haplotype_reconstruction(geno_data, error_rate = 0.05, r = 0.02)
print(reconstructed_geno_custom_r)
}
}
