% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_install.R
\name{load_or_install_cran}
\alias{load_or_install_cran}
\title{Load or install a CRAN package}
\usage{
load_or_install_cran(pkg)
}
\arguments{
\item{pkg}{Character. Package name.}
}
\value{
Loads the package. Installs it if not already installed.
}
\description{
Load or install a CRAN package
}
