% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_install.R, R/load_or_install_github.R
\name{load_or_install_github}
\alias{load_or_install_github}
\title{Load or install a GitHub package}
\usage{
load_or_install_github(pkg, repo)

load_or_install_github(pkg, repo)
}
\arguments{
\item{pkg}{A character string specifying the name of the package to be loaded or installed.}

\item{repo}{A character string specifying the GitHub repository in the form "user/repo" from which to install the package.}
}
\value{
Loads the package. Installs it from GitHub if needed.

This function does not return a value. It is called for its side effect of loading or installing the package.
}
\description{
This function checks if a specified package is installed. If the package is not installed,
it installs the package from GitHub using the \code{remotes} package, and then loads it into the R session.
}
\details{
The function first checks if the package is available in the current session using \code{requireNamespace}.
If the package is not installed, it installs it from the specified GitHub repository using \code{remotes::install_github},
and then loads the package using \code{library}.
}
\examples{
\dontrun{
load_or_install_github("geneticMapR", "vegaalfaro/geneticMapR")
load_or_install_github("dplyr", "tidyverse/dplyr")
}

}
